Add delay to catwalk construction

+++ code/game/turfs/space/space.dm
@@ -44,8 +44,11 @@
 			if(R.use(1))
 				user << "<span class='notice'>Constructing catwalk...</span>"
 				playsound(src, 'sound/weapons/Genhit.ogg', 50, 1)
-				qdel(L)
-				ReplaceWithCatwalk()
+				if(do_after(user, 10, 5, 0))
+					qdel(L)
+					ReplaceWithCatwalk()
+					user << "<span class='notice'>Catwalk constructed</span>"
+
 			else
 				user << "<span class='warning'>You need two rods to build a catwalk.</span>"
 			return
